feat(simd): implement optimizations and enhance documentation
- Implemented SSE4.1 and WASM SIMD128 optimizations for hot paths.
- Improved CI/CD pipeline with GitHub Actions.
- Created ARCHITECTURE.md with deep dive into SIMD optimizations and design.
- Added performance benchmarks and reproduction steps.
- Enhanced README.md with status badges and navigation.

+++ crates/core/src/types.rs
@@ -319,9 +319,9 @@ pub struct ARPattHandle {
     pub patt_num: i32,
     pub patt_num_max: i32,
     pub pattf: Vec<i32>,
-    pub patt: Vec<Vec<i32>>,
+    pub patt: Vec<Vec<i16>>,
     pub pattpow: Vec<ARdouble>,
-    pub patt_bw: Vec<Vec<i32>>,
+    pub patt_bw: Vec<Vec<i16>>,
     pub pattpow_bw: Vec<ARdouble>,
     pub patt_size: i32,
 }
